+++ lib/externalApis/tenderdash/waitForHeightFactory.js
@@ -1,4 +1,4 @@
-const BlockchainListener = require('./blockchainListener/BlockchainListener');
+const BlockchainListener = require('./BlockchainListener');
 
 /**
  * @param {BlockchainListener} blockchainListener
