+++ lib/externalApis/tenderdash/waitForTransactionToBeProvable/getExistingTransactionResult.js
