+++ lib/externalApis/tenderdash/waitForTransactionToBeProvable/transactionResult/AbstractTransactionResult.js
