+++ lib/externalApis/tenderdash/waitForTransactionToBeProvable/transactionResult/TransactionErrorResult.js
