+++ lib/externalApis/tenderdash/waitForTransactionToBeProvable/transactionResult/TransactionOkResult.js
