+++ lib/externalApis/tenderdash/waitForTransactionToBeProvable/waitForTransactionResult.js
