+++ lib/externalApis/tenderdash/waitForTransactionToBeProvable/waitForTransactionToBeProvableFactory.js
@@ -1,4 +1,4 @@
-const TransactionWaitPeriodExceededError = require('../../../../errors/TransactionWaitPeriodExceededError');
+const TransactionWaitPeriodExceededError = require('../../../errors/TransactionWaitPeriodExceededError');
 const TransactionOkResult = require('./transactionResult/TransactionOkResult');
 
 /**
