+++ lib/grpcServer/handlers/platform/platformHandlersFactory.js
@@ -68,10 +68,10 @@ const waitForStateTransitionResultHandlerFactory = require(
 );
 
 const fetchProofForStateTransitionFactory = require('../../../externalApis/drive/fetchProofForStateTransitionFactory');
-const waitForTransactionToBeProvableFactory = require('../../../externalApis/tenderdash/blockchainListener/waitForTransactionToBeProvable/waitForTransactionToBeProvableFactory');
-const waitForTransactionResult = require('../../../externalApis/tenderdash/blockchainListener/waitForTransactionToBeProvable/waitForTransactionResult');
+const waitForTransactionToBeProvableFactory = require('../../../externalApis/tenderdash/waitForTransactionToBeProvable/waitForTransactionToBeProvableFactory');
+const waitForTransactionResult = require('../../../externalApis/tenderdash/waitForTransactionToBeProvable/waitForTransactionResult');
 const waitForHeightFactory = require('../../../externalApis/tenderdash/waitForHeightFactory');
-const getExistingTransactionResultFactory = require('../../../externalApis/tenderdash/blockchainListener/waitForTransactionToBeProvable/getExistingTransactionResult');
+const getExistingTransactionResultFactory = require('../../../externalApis/tenderdash/waitForTransactionToBeProvable/getExistingTransactionResult');
 
 /**
  * @param {jaysonClient} rpcClient
