+++ lib/grpcServer/handlers/platform/waitForStateTransitionResultHandlerFactory.js
@@ -18,7 +18,7 @@ const {
 const cbor = require('cbor');
 
 const TransactionWaitPeriodExceededError = require('../../../errors/TransactionWaitPeriodExceededError');
-const TransactionErrorResult = require('../../../externalApis/tenderdash/blockchainListener/waitForTransactionToBeProvable/transactionResult/TransactionErrorResult');
+const TransactionErrorResult = require('../../../externalApis/tenderdash/waitForTransactionToBeProvable/transactionResult/TransactionErrorResult');
 
 /**
  *
